mv get_position_ids_from_cu_seq_lens to flash util
